Add recipe for converting `JUnit#assertThrows` to `AssertJ#AssertExceptionType`. Closes #171 (#172)
@timtebeek Thanks for the suggestion!
I have added a new recipe for converting Junit#assertThrows to AssertJ#AssertExceptionType. In case you're interested there is some added complexity of maintaining the AssertJ shim used by the JavaTemplate. For that, we have a template-generator project which use for generating JavaTemplate shims, some additional formatting is required but it generally saves a lot of work.