Don't remove `mockStatic()` calls in resources of try-with-resources
This commit does not yet close the issue, as there are other cases, where this can potentially go wrong.

It also appears strange that the recipe matches on Mockito's `mockStatic()` rather than on PowerMock's `mockStatic()`. Therefore, the recipe is for now excluded from `Mockito1to3Migration`, as this is part of `JUnit4to5Migration` which is widely used.

Issue: #360