Create a mockito upgrade recipe that just does 4 -> 5, so it can be included in Spring 3.1 upgrade.
Both "Migrate to Spring Boot 2.4" and "Mockito 5.x upgrade" include "Mockito 4.x upgrade"
So if we add "Mockito 5.x upgrade" into "Migrate to Spring Boot 3.1" then "Mockito 4.x upgrade", and all its constituent recipes, will be run twice. The performance cost of this is not dire, but the spring boot migration is already a very large recipe with ~2k steps.