Migrate Hamcrest to JUnit 5 (#343)
* WIP: Add recipe for migration from Hamcrest

Work in progress implementation of a recipe which migrates Hamcrest test matchers to JUnit5 test assertions.

Signed-off-by: matus.matok <matus.matok@pantheon.tech>

* Add missing license headers

* Resolve some of the test issues

* Fix test import

* Add proto implementation for assertEquals

Added a prototype-y implementation of translation from hamcrest's equalTo to JUnit5's assertEquals. Should be easy to add more of the simple hamcrest matchers to this implementation.

\TODO the import is not being added

Signed-off-by: matus.matok <matus.matok@pantheon.tech>

* Use static import and #{any(java.lang.Object)} to fix test

* Adapt to main

Another iteration of the prototype. Adapted it to be
based on the up-to-date main.
Added a proposal of how similar simple matchers could be
translated to junit assertion methods.

Signed-off-by: matus.matok <matus.matok@pantheon.tech>

* Add more simple matcher-to-method translations

Added translations for matchers closeTo, containsString, empty, emptyArray, emptyIterable, emptyCollectionOf, emptyIterableOf, endsWith.
Tests need to be added for each matcher.

Signed-off-by: matus.matok <matus.matok@pantheon.tech>

* Add more simple matcher-to-method translations

Added all the simple matcher-to-method translations.
Tests need to be added for each matcher.

Signed-off-by: matus.matok <matus.matok@pantheon.tech>

* Add tests

Added a bunch of tests, to verify the correctness of each matcher-to-assertion case.

Signed-off-by: matus.matok <matus.matok@pantheon.tech>

* Finalise the pull request

Added all the necessary unit tests and polished out the implementation.

Signed-off-by: matus.matok <matus.matok@pantheon.tech>

* Add required license header

* Move classes to align with the Hamcrest to AssertJ implementation

* Consistently use `class Test` to avoid conflicts with `@Test`

* Refactored and split HamcrestMatcherToJUnit5 recipe

Refactored HamcrestMatcherToJUnit5 recipe, so now the whole translation is stored in one place, not scatter amongst three methods. Given the existence of RemoveIsMatcher Recipe, this recipe relies that it will never encounter is() matcher.
In similar fashion as RemoveIsMatcher, RemoveNotMatcher was added, which does exactly the same as RemoveIsMatcher, but it also stores the logical context for the nested matcher (so that it knows it was negated) in execution context.
Matchers instanceOf and isA were difficult to handle within the HamcrestMatcherToJUnit5 recipe, therefore these cases were moved to a newly added HamcrestInstanceOfToJUnit5 recipe.

Signed-off-by: matus.matok <matus.matok@pantheon.tech>

* Add license headers

Forgot, added now

Signed-off-by: matus.matok <matus.matok@pantheon.tech>

* Apply suggestions from code review

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

* Fix compilation

* Update description to use JUnit

* Apply suggestions from code review

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

* Format tests

* Drop RemoveNotMatcher recipe; retain visitor only

To avoid misuse

* Drop AssertThatBooleanToJUnit5; replace with declarative recipes

* Apply suggestions from code review

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

* Limit execution through preconditions

* Add preconditions to HamcrestMatcherToJUnit5

* Minor polish

* Extract and reuse `ConsistentHamcrestMatcherImports`

* Polish recipe display name and description

---------

Signed-off-by: matus.matok <matus.matok@pantheon.tech>
Co-authored-by: Tim te Beek <tim@moderne.io>
Co-authored-by: Tim te Beek <timtebeek@gmail.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>